Move the function for detecting peaks into a separate module
